Error Handling Refactor with Mixin
